+++ CHANGELOG.md
@@ -1,7 +1,10 @@
 ### Upcoming
+- Changed the signature of `System.start`.  `@co` usage is unaffected.
+- Added `cancelIfCoroutineStarted` to implement mutually exclusive coroutines.
+- Made sure to reset `accessRecentlyDeletedData` between coroutines.
 
 ### 0.9.4
-- Fix the perf build, which got broken in 0.9.2.
+- Fixed the perf build, which got broken in 0.9.2.
 
 ### 0.9.3
 - Put field types directly off the `field` decorator, so you can just do `@field.int32` instead of `@field(Type.int32)`.  Less imports, less typing, more readable -- a win all around!  You still need to use the long form if you want to specify any field options, though.